+++ test/01_removing_pool.js
