+++ utils/constants.js
@@ -1,8 +1,9 @@
 'use strict';
-const BLOCK_REWARD_ADDRESS = require("../posdao-contracts/spec").engine.authorityRound.params.blockRewardContractAddress;
-const VALIDATOR_SET_ADDRESS = require("../posdao-contracts/spec").engine.authorityRound.params.validators.multi[0].contract;
+const specParams = require("../posdao-contracts/spec").engine.authorityRound.params;
+const BLOCK_REWARD_ADDRESS = specParams.blockRewardContractAddress;
+const VALIDATOR_SET_ADDRESS = specParams.validators.multi[0].contract;
+const RANDOM_AURA_ADDRESS = specParams.randomnessContractAddress;
 const STAKING_CONTRACT_ADDRESS = "0x1100000000000000000000000000000000000001";
-const RANDOM_AURA_ADDRESS = "0x3000000000000000000000000000000000000001";
 
 module.exports = {
   OWNER: "0x32e4e4c7c5d1cea5db5f9202a9e4d99e56c91a24",
